0.1 resolve another comment

+++ FinanceTracker/FinanceTracker.Contracts/Models/Transactions/GetTransaction/GetTransactionDto.cs
@@ -5,5 +5,5 @@ namespace FinanceTracker.Contracts.Models.Transactions.GetTransaction;
 public class GetTransactionDto(Transaction transaction)
 {
     public int Id { get; init; } = transaction.Id;
-    public double Amount { get; init; } = transaction.Amount;
+    public decimal Amount { get; init; } = transaction.Amount;
 }
